Add class DatasetCacher.
Caches the output of a whole dataset to disk.

If used with batch_size > 1 and no collate function, it iterates over
all samples individually and caches the list to disk.

If used with batcH_size <= 1 and a collate function, the tensor with
first dimension (batch_size, dim1,  ... ) is cached to disk instead. The
cache can be used by a DataLoader with automatic batching disabled to
avoid having to run the collate_fn later on every epoch.